+++ packages/cloud-cognitive/src/components/CreateFullPage/constants.js
@@ -1,0 +1,2 @@
+export const CREATE_FULL_PAGE_STEP = 'CreateFullPageStep';
+export const CREATE_FULL_PAGE_SECTION = 'CreateFullPageSection';