+++ packages/cloud-cognitive/src/components/index.js
@@ -10,6 +10,11 @@ export { ActionBarItem } from './ActionBar';
 export { ButtonMenu, ButtonMenuItem } from './ButtonMenu';
 export { ComboButton, ComboButtonItem } from './ComboButton';
 export { ContextHeader } from './ContextHeader';
+export {
+  CreateFullPage,
+  CreateFullPageStep,
+  CreateFullPageSection,
+} from './CreateFullPage';
 export { CreateModal } from './CreateModal';
 export { CreateSidePanel } from './CreateSidePanel';
 export { CreateTearsheet, CreateTearsheetStep } from './CreateTearsheet';
@@ -34,6 +39,7 @@ export { RemoveModal } from './RemoveModal';
 export { Saving } from './Saving';
 export { SidePanel } from './SidePanel';
 export { StatusIcon } from './StatusIcon';
+export { TagSet } from './TagSet';
 export { Tearsheet, TearsheetNarrow } from './Tearsheet';
 export { UserProfileImage } from './UserProfileImage';
 export { WebTerminal } from './WebTerminal';